+++ guillotina/tests/test_attachment.py
@@ -1,4 +1,4 @@
-from guillotina.behaviors.attachment import IAttachment
+from guillotina.behaviors.attachment import IAttachment, IMultiAttachment
 from guillotina.component import get_multi_adapter
 from guillotina.interfaces import IFileManager
 from guillotina.tests import utils
@@ -37,6 +37,64 @@ async def test_create_content_with_behavior(container_requester):
         )
         assert status == 200
         assert len(response) == (1024 * 1024 * 4)
+
+
+async def test_multi_upload(container_requester):
+    async with container_requester as requester:
+        response, status = await requester(
+            'POST',
+            '/db/guillotina/',
+            data=json.dumps({
+                '@type': 'Item',
+                '@behaviors': [IMultiAttachment.__identifier__],
+                'id': 'foobar'
+            })
+        )
+        assert status == 201
+
+        response, status = await requester(
+            'PATCH',
+            '/db/guillotina/foobar/@upload/files/key1',
+            data=b'X' * 1024 * 1024 * 10,
+            headers={
+                'x-upload-size': str(1024 * 1024 * 10)
+            }
+        )
+        assert status == 200
+
+        response, status = await requester(
+            'PATCH',
+            '/db/guillotina/foobar/@upload/files/key2',
+            data=b'Y' * 1024 * 1024 * 10,
+            headers={
+                'x-upload-size': str(1024 * 1024 * 10)
+            }
+        )
+        assert status == 200
+
+        response, status = await requester(
+            'GET',
+            '/db/guillotina/foobar/@download/files/key1'
+        )
+        assert status == 200
+        assert response == b'X' * 1024 * 1024 * 10
+
+        response, status = await requester(
+            'GET',
+            '/db/guillotina/foobar/@download/files/key2'
+        )
+        assert status == 200
+        assert response == b'Y' * 1024 * 1024 * 10
+
+        request = utils.get_mocked_request(requester.db)
+        root = await utils.get_root(request)
+        async with managed_transaction(request=request, abort_when_done=True):
+            container = await root.async_get('guillotina')
+            obj = await container.async_get('foobar')
+            behavior = IMultiAttachment(obj)
+            await behavior.load()
+            assert behavior.files['key1']._blob.chunks == 2
+            assert behavior.files['key2']._blob.chunks == 2
 
 
 async def test_large_upload_chunks(container_requester):