+++ CHANGELOG.rst
@@ -3,6 +3,8 @@
 
 - Implement multi attachments
   [masipcat]
+- Allow to login on IApplication.
+  [bloodbare]
 
 
 4.4.4 (2019-01-11)